2.Junior/2.In_and_out_put/src/main/java/ru/job4j/io/Search.java
package ru.job4j.io;

import java.io.IOException;
import java.nio.file.*;
import java.nio.file.attribute.BasicFileAttributes;
import java.util.ArrayList;
import java.util.List;

import static java.nio.file.FileVisitResult.CONTINUE;

/**
 * Разработайте программу Search, которая будет выводит файлы только с определенным расширением.
 * Created by Roman on 09.04.2020
 */
public class Search {
    public static void main(String[] args) throws IOException {
        Path start = Paths.get(".");
        Files.walkFileTree(start, new PrintFiles());
    }

    public static List<String> search(Path root, String ext) throws IOException {
        List<String> result = new ArrayList<>();
        Files.walkFileTree(root, new SimpleFileVisitor<>() {
            @Override
            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) {
                String fileName = file.toString();
                if (fileName.endsWith(ext)) {
                    result.add(fileName);
                }
                return CONTINUE;
            }
        });
        return result;
    }
}
